+++ packages/core/src/zig/buffer.zig
@@ -85,7 +85,7 @@ pub const Cell = struct {
     char: u32,
     fg: RGBA,
     bg: RGBA,
-    attributes: u8,
+    attributes: u32,
 };
 
 fn isRGBAWithAlpha(color: RGBA) bool {
@@ -124,7 +124,7 @@ pub const OptimizedBuffer = struct {
         char: []u32,
         fg: []RGBA,
         bg: []RGBA,
-        attributes: []u8,
+        attributes: []u32,
     },
     width: u32,
     height: u32,
@@ -166,7 +166,7 @@ pub const OptimizedBuffer = struct {
                 .char = allocator.alloc(u32, size) catch return BufferError.OutOfMemory,
                 .fg = allocator.alloc(RGBA, size) catch return BufferError.OutOfMemory,
                 .bg = allocator.alloc(RGBA, size) catch return BufferError.OutOfMemory,
-                .attributes = allocator.alloc(u8, size) catch return BufferError.OutOfMemory,
+                .attributes = allocator.alloc(u32, size) catch return BufferError.OutOfMemory,
             },
             .width = width,
             .height = height,
@@ -199,7 +199,7 @@ pub const OptimizedBuffer = struct {
         return self.buffer.bg.ptr;
     }
 
-    pub fn getAttributesPtr(self: *OptimizedBuffer) [*]u8 {
+    pub fn getAttributesPtr(self: *OptimizedBuffer) [*]u32 {
         return self.buffer.attributes.ptr;
     }
 
@@ -574,7 +574,7 @@ pub const OptimizedBuffer = struct {
         char: u32,
         fg: RGBA,
         bg: RGBA,
-        attributes: u8,
+        attributes: u32,
     ) !void {
         if (!self.isPointInScissor(@intCast(x), @intCast(y))) return;
         const overlayCell = Cell{ .char = char, .fg = fg, .bg = bg, .attributes = attributes };
@@ -594,7 +594,7 @@ pub const OptimizedBuffer = struct {
         char: u32,
         fg: RGBA,
         bg: RGBA,
-        attributes: u8,
+        attributes: u32,
     ) !void {
         if (!self.isPointInScissor(@intCast(x), @intCast(y))) return;
         const overlayCell = Cell{ .char = char, .fg = fg, .bg = bg, .attributes = attributes };
@@ -618,7 +618,7 @@ pub const OptimizedBuffer = struct {
         y: u32,
         fg: RGBA,
         bg: RGBA,
-        attributes: u8,
+        attributes: u32,
     ) !void {
         if (!self.isPointInScissor(@intCast(x), @intCast(y))) return;
 
@@ -701,7 +701,7 @@ pub const OptimizedBuffer = struct {
         y: u32,
         fg: RGBA,
         bg: ?RGBA,
-        attributes: u8,
+        attributes: u32,
     ) BufferError!void {
         if (x >= self.width or y >= self.height) return;
         if (text.len == 0) return;
